@@ -1,5 +1,14 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
+import React from "react";
+import ReactDOM from "react-dom/client";
+import App from "./App";
+import { FirebaseContext } from "./Store/Context";
+import { auth } from "./Firebase/config";
+import Context from "./Store/Context";
 
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.createRoot(document.getElementById("root")).render(
+  <FirebaseContext.Provider value={{ auth }}>
+    <Context>
+      <App />
+    </Context>
+  </FirebaseContext.Provider>
+);
