@@ -1,17 +1,44 @@
-import React from 'react';
+import React, { useState , useContext} from 'react';
+// import { FirebaseContext } from '../../Store/Context';
+import { signInWithEmailAndPassword } from 'firebase/auth'
+import { auth } from '../../Firebase/config';
+import { useNavigate } from 'react-router-dom/dist';
 
 import Logo from '../../olx-logo.png';
 import './Login.css';
+import Header from '../Header/Header';
+
 
 function Login() {
+  const [email,setEmail] = useState('')
+  const [password,setPassword] = useState('')
+  // const { firebase } = useContext(FirebaseContext);
+  const navigate = useNavigate();
+
+  const loginHandler = (e)=>{
+    e.preventDefault()
+    signInWithEmailAndPassword(auth,email,password).then((result)=>{
+      navigate('/')
+      
+    }).catch((error)=>{
+      alert(error.message)
+    })
+  }
   return (
-    <div>
+    
+    <div >
+      <Header/>
       <div className="loginParentDiv">
         <img width="200px" height="200px" src={Logo}></img>
-        <form>
-          <label htmlFor="fname">Email</label>
+        <form onSubmit={loginHandler}>
+          <label htmlFor="fname">Enter your email to Login.</label>
           <br />
           <input
+          value={email}
+          onChange={(e)=>{
+            setEmail(e.target.value)
+          }}
+          placeholder='Email Address'
             className="input"
             type="email"
             id="fname"
@@ -19,9 +46,14 @@ function Login() {
             defaultValue="John"
           />
           <br />
-          <label htmlFor="lname">Password</label>
+          <label htmlFor="lname"></label>
           <br />
           <input
+          placeholder='Password'
+          value={password}
+          onChange={(e)=>{
+            setPassword(e.target.value)
+          }}
             className="input"
             type="password"
             id="lname"
@@ -32,9 +64,15 @@ function Login() {
           <br />
           <button>Login</button>
         </form>
-        <a>Signup</a>
+        <p>
+          Don't have an account ? <a onClick={()=>{
+            navigate('/signup')
+          }}>Sign Up</a>
+        </p>
       </div>
+      
     </div>
+    
   );
 }
 
