@@ -1,9 +1,29 @@
-import React from 'react';
+import React, { useContext, useEffect, useState } from "react";
+import { Link } from "react-router-dom/dist";
+import { postContext } from "../../Store/PostContext";
+import { firestore } from "../../Firebase/config";
 
-import Heart from '../../assets/Heart';
-import './Post.css';
+import { collection, getDocs } from "firebase/firestore";
+
+import Heart from "../../assets/Heart";
+import "./Post.css";
 
 function Posts() {
+  const [products, setProducts] = useState([]);
+  const { setPostDetails } = useContext(postContext);
+  useEffect(() => {
+    getDocs(collection(firestore, "products"))
+      .then((querySnapshot) => {
+        const allProducts = querySnapshot.docs.map((doc) => ({
+          ...doc.data(),
+          id: doc.id,
+        }));
+        setProducts(allProducts);
+      })
+      .catch((error) => {
+        console.error("Error fetching products:", error);
+      });
+  }, []);
 
   return (
     <div className="postParentDiv">
@@ -13,47 +33,68 @@ function Posts() {
           <span>View more</span>
         </div>
         <div className="cards">
-          <div
-            className="card"
-          >
-            <div className="favorite">
-              <Heart></Heart>
-            </div>
-            <div className="image">
-              <img src="../../../Images/R15V3.jpg" alt="" />
-            </div>
-            <div className="content">
-              <p className="rate">&#x20B9; 250000</p>
-              <span className="kilometer">Two Wheeler</span>
-              <p className="name"> YAMAHA R15V3</p>
-            </div>
-            <div className="date">
-              <span>Tue May 04 2021</span>
-            </div>
-          </div>
+          {products.map((product) => {
+            return (
+              <Link to={"/viewPost"}>
+                <div
+                  className="card"
+                  onClick={() => {
+                    setPostDetails(product);
+                  }}
+                  key={product.url}
+                >
+                  <div className="favorite">
+                    <Heart></Heart>
+                  </div>
+                  <div className="image">
+                    <img src={product.url} alt="" />
+                  </div>
+                  <div className="content">
+                    <p className="rate">&#x20B9; {product.price}</p>
+                    <span className="kilometer">{product.name}</span>
+                    <p className="name"> {product.category}</p>
+                  </div>
+                  <div className="date">
+                    <span>{product.createdAt}</span>
+                  </div>
+                </div>
+              </Link>
+            );
+          })}
         </div>
       </div>
       <div className="recommendations">
         <div className="heading">
           <span>Fresh recommendations</span>
         </div>
-        <div className="cards">
-          <div className="card">
-            <div className="favorite">
-              <Heart></Heart>
-            </div>
-            <div className="image">
-              <img src="../../../Images/R15V3.jpg" alt="" />
-            </div>
-            <div className="content">
-              <p className="rate">&#x20B9; 250000</p>
-              <span className="kilometer">Two Wheeler</span>
-              <p className="name"> YAMAHA R15V3</p>
-            </div>
-            <div className="date">
-              <span>10/5/2021</span>
-            </div>
-          </div>
+        <div className="freshcards">
+          {products.map((product) => {
+            return (
+              <Link to={"/viewPost"}>
+                <div
+                  className="card"
+                  onClick={() => {
+                    setPostDetails(product);
+                  }}
+                >
+                  <div className="favorite">
+                    <Heart></Heart>
+                  </div>
+                  <div className="image">
+                    <img src={product.url} alt="" />
+                  </div>
+                  <div className="content">
+                    <p className="rate">&#x20B9; {product.price}</p>
+                    <span className="kilometer">{product.name}</span>
+                    <p className="name"> {product.category}</p>
+                  </div>
+                  <div className="date">
+                    <span>{product.createdAt}</span>
+                  </div>
+                </div>
+              </Link>
+            );
+          })}
         </div>
       </div>
     </div>
